Always set $VERBOSE to true and enable logging of invalid refs.
