+++ asciidoctorj-core/src/main/resources/org/asciidoctor/internal/asciidoctorclass.rb
@@ -1,3 +1,6 @@
+# Set the global variable VERBOSE to true to get invalid refs into the log
+$VERBOSE=true
+
 module AsciidoctorJ
     include_package 'org.asciidoctor'
     module Extensions
