+++ asciidoctorj-core/src/test/resources/documentwithinvalidrefs.adoc
@@ -1,0 +1,3 @@
+= Test
+
+This is an <<invalidref>>
